+++ pytest_container/inspect.py
@@ -2,6 +2,7 @@
 :command:`$runtime inspect $ctr_id`.
 
 """
+
 import enum
 import socket
 from dataclasses import dataclass