+++ tests/test_lock.py
@@ -4,6 +4,7 @@ from time import sleep
 
 from _pytest.config import Config
 from filelock import FileLock
+
 from pytest_container import Container
 from pytest_container import DerivedContainer
 from pytest_container.container import ContainerData