+++ tests/test_runtime.py
@@ -1,13 +1,20 @@
 # pylint: disable=missing-function-docstring,missing-module-docstring
 import os
 from pathlib import Path
+from typing import Callable
+from typing import Type
+from typing import Union
 from unittest.mock import patch
 
 import pytest
+
+from pytest_container.runtime import LOCALHOST
 from pytest_container.runtime import DockerRuntime
-from pytest_container.runtime import get_selected_runtime
 from pytest_container.runtime import OciRuntimeBase
 from pytest_container.runtime import PodmanRuntime
+from pytest_container.runtime import Version
+from pytest_container.runtime import _get_buildah_version
+from pytest_container.runtime import get_selected_runtime
 
 
 @pytest.fixture
@@ -18,6 +25,58 @@ def container_runtime_envvar(request):
         clear=True,
     ):
         yield
+
+
+# pylint: disable-next=unused-argument
+def _mock_run_success(*args, **kwargs):
+    class Succeeded:
+        """Class that mocks the returned object of `testinfra`'s `run`."""
+
+        @property
+        def succeeded(self) -> bool:
+            return True
+
+        @property
+        def rc(self) -> int:
+            return 0
+
+    return Succeeded()
+
+
+def generate_mock_fail(*, rc: int = 1, stderr: str = "failure!!"):
+    # pylint: disable-next=unused-argument
+    def mock_run_fail(cmd: str):
+        class Failure:
+            """Class that mocks the returned object of `testinfra`'s `run`."""
+
+            @property
+            def succeeded(self) -> bool:
+                return False
+
+            @property
+            def rc(self) -> int:
+                return rc
+
+            @property
+            def stderr(self) -> str:
+                return stderr
+
+        return Failure()
+
+    return mock_run_fail
+
+
+def _create_mock_exists(
+    podman_should_exist: bool, docker_should_exist: bool
+) -> Callable[[str], bool]:
+    def exists(prog: str) -> bool:
+        if prog == "podman" and podman_should_exist:
+            return True
+        if prog == "docker" and docker_should_exist:
+            return True
+        return False
+
+    return exists
 
 
 @pytest.mark.parametrize(
@@ -35,8 +94,73 @@ def test_runtime_selection(
     # pylint: disable-next=redefined-outer-name,unused-argument
     container_runtime_envvar: None,
     runtime: OciRuntimeBase,
+    monkeypatch: pytest.MonkeyPatch,
 ):
+    monkeypatch.setattr(LOCALHOST, "run", _mock_run_success)
+    monkeypatch.setattr(LOCALHOST, "exists", _create_mock_exists(True, True))
+
     assert get_selected_runtime() == runtime
+
+
+@pytest.mark.parametrize("runtime", ("podman", "docker"))
+def test_value_err_when_docker_and_podman_missing(
+    runtime: str, monkeypatch: pytest.MonkeyPatch
+) -> None:
+    monkeypatch.setenv("CONTAINER_RUNTIME", runtime)
+    monkeypatch.setattr(LOCALHOST, "exists", _create_mock_exists(False, False))
+    with pytest.raises(ValueError) as val_err_ctx:
+        get_selected_runtime()
+
+    assert f"Selected runtime {runtime} does not exist on the system" in str(
+        val_err_ctx.value
+    )
+
+
+@pytest.mark.parametrize(
+    "cls, name", ((PodmanRuntime, "podman"), (DockerRuntime, "docker"))
+)
+def test_runtime_construction_fails_if_ps_fails(
+    cls: Union[Type[PodmanRuntime], Type[DockerRuntime]],
+    name: str,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    stderr = "container runtime failed"
+    monkeypatch.setattr(LOCALHOST, "run", generate_mock_fail(stderr=stderr))
+    with pytest.raises(RuntimeError) as rt_err_ctx:
+        cls()
+
+    assert f"`{name} ps` failed with {stderr}" in str(rt_err_ctx.value)
+
+
+@pytest.mark.parametrize(
+    "version_str, expected_version",
+    (
+        ("1.38.0", Version(1, 38, 0)),
+        ("1.25.3", Version(1, 25, 3)),
+    ),
+)
+def test_buildah_version_parsing(
+    version_str: str,
+    expected_version: Version,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    monkeypatch.setattr(
+        LOCALHOST, "check_output", lambda _: f"buildah version {version_str}"
+    )
+
+    assert _get_buildah_version() == expected_version
+
+
+def test_get_buildah_version_fails_on_unexpected_stdout(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    monkeypatch.setattr(LOCALHOST, "check_output", lambda _: "foobar")
+    with pytest.raises(RuntimeError) as rt_err_ctx:
+        _get_buildah_version()
+
+    assert "Could not decode the buildah version from 'foobar'" in str(
+        rt_err_ctx.value
+    )
 
 
 @pytest.mark.parametrize(