+++ tests/test_volumes.py
@@ -5,14 +5,15 @@ from os.path import join
 from typing import List
 
 import pytest
+
 from pytest_container.container import BindMount
 from pytest_container.container import BindMountCreator
 from pytest_container.container import ContainerData
 from pytest_container.container import ContainerVolume
 from pytest_container.container import ContainerVolumeBase
 from pytest_container.container import DerivedContainer
-from pytest_container.container import get_volume_creator
 from pytest_container.container import VolumeFlag
+from pytest_container.container import get_volume_creator
 from pytest_container.runtime import LOCALHOST
 from pytest_container.runtime import OciRuntimeBase